requirements.txt
selenium==4.15.2
flask==2.0.1
python-telegram-bot==13.7
psutil==5.8.0 